requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
deprecation==2.1.0
fastapi==0.115.12
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.20.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
newsapi-python==0.2.7
numpy==2.3.0
packaging==25.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyTelegramBotAPI==4.27.0
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
realtime==2.4.3
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlite-vec==0.1.6
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supafunc==0.9.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==14.2
yarl==1.20.1
